refactor(rag): promote extraction.py to extraction/ package
extraction.py grew to handle PDF, DOCX, HTML, and plain-text formats
in a single flat file, making it difficult to add or modify an extractor
without touching unrelated code. Promote it to a package with one module
per format:
- base.py   — shared abstract base and protocol
- pdf.py    — PDF extraction logic
- docx.py   — DOCX extraction logic
- html.py   — HTML extraction logic
- txt.py    — plain-text extraction logic
- utils.py  — shared helpers
- __init__.py — public re-exports; import surface is unchanged for callers
