+++ app/rag/extraction/txt.py
@@ -1,0 +1,15 @@
+"""Plain-text and Markdown extractor."""
+
+from app.rag.extraction.base import BaseExtractor
+from app.rag.types import DocType, ExtractionResult
+
+
+class TXTExtractor(BaseExtractor):
+    """Extracts text from plain-text and Markdown files."""
+
+    def extract(self, data: bytes, source_name: str) -> ExtractionResult:
+        return ExtractionResult(
+            markdown=data.decode("utf-8", errors="replace"),
+            doc_type=DocType.TXT,
+            source_name=source_name,
+        )
